@@ -1,44 +1,61 @@
 {
   "name": "passvault",
-  "version": "1.0.0",
   "main": "expo-router/entry",
+  "version": "1.0.0",
   "scripts": {
     "start": "expo start",
-    "android": "expo start --android",
-    "ios": "expo start --ios",
+    "android": "expo run:android",
+    "ios": "expo run:ios",
     "web": "expo start --web",
-    "format": "prettier . -w"
+    "test": "jest --watchAll"
+  },
+  "jest": {
+    "preset": "jest-expo"
   },
   "dependencies": {
-    "@babel/runtime": "^7.25.7",
-    "@clerk/clerk-expo": "^2.2.30",
-    "@clerk/nextjs": "^6.1.0",
-    "@react-native/assets-registry": "^0.74.88",
-    "@types/react": "~18.2.79",
+    "@expo/vector-icons": "^14.0.2",
+    "@react-native-community/slider": "4.5.5",
+    "@react-navigation/native": "^7.0.0",
+    "@tanstack/react-query": "^5.60.5",
+    "@types/react-native-crypto-js": "^1.0.3",
     "axios": "^1.7.7",
-    "expo": "~51.0.37",
-    "expo-asset": "~10.0.10",
-    "expo-constants": "~16.0.2",
-    "expo-font": "~12.0.10",
-    "expo-linking": "~6.3.1",
-    "expo-router": "~3.5.23",
-    "expo-secure-store": "^13.0.2",
-    "expo-status-bar": "~1.12.1",
-    "nativewind": "^2.0.11",
-    "prettier": "^3.3.3",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "react-native": "0.74.5",
+    "expo": "~52.0.7",
+    "expo-checkbox": "~4.0.0",
+    "expo-clipboard": "~7.0.0",
+    "expo-crypto": "~14.0.1",
+    "expo-dev-client": "~5.0.3",
+    "expo-font": "~13.0.1",
+    "expo-linking": "~7.0.2",
+    "expo-navigation-bar": "~4.0.5",
+    "expo-router": "~4.0.6",
+    "expo-secure-store": "~14.0.0",
+    "expo-splash-screen": "~0.29.11",
+    "expo-status-bar": "~2.0.0",
+    "expo-system-ui": "~4.0.3",
+    "expo-updates": "~0.26.7",
+    "expo-web-browser": "~14.0.1",
+    "iconsax-react-native": "^0.0.8",
+    "react": "18.3.1",
+    "react-dom": "18.3.1",
+    "react-native": "0.76.2",
     "react-native-crypto-js": "^1.0.0",
-    "react-native-safe-area-context": "4.10.5",
-    "react-native-screens": "~3.31.1",
-    "react-native-web": "~0.19.12",
-    "typescript": "~5.3.3"
+    "react-native-element-dropdown": "^2.12.2",
+    "react-native-react-query-devtools": "^1.3.10",
+    "react-native-reanimated": "~3.16.1",
+    "react-native-safe-area-context": "4.12.0",
+    "react-native-screens": "~4.0.0",
+    "react-native-svg": "^15.9.0",
+    "react-native-toast-message": "^2.2.1",
+    "react-native-web": "~0.19.13",
+    "zustand": "^5.0.1"
   },
   "devDependencies": {
-    "@babel/core": "^7.25.7",
-    "@types/node": "^20.16.10",
-    "tailwindcss": "3.3.2"
+    "@babel/core": "^7.25.2",
+    "@types/react": "~18.3.12",
+    "jest": "^29.2.1",
+    "jest-expo": "~52.0.2",
+    "react-test-renderer": "18.3.1",
+    "typescript": "~5.3.3"
   },
   "private": true
 }
