@@ -1,16 +1,22 @@
-const CryptoJS = require('react-native-crypto-js')
+import * as CryptoJS from "react-native-crypto-js";
 
-const secretKey = process.env.EXPO_PUBLIC_ENCRYPTION_SECRET_KEY;
-console.log(secretKey)
+const secretKey = process.env.EXPO_PUBLIC_ENCRYPTION_SECRET_KEY as string;
+console.log(secretKey);
 
 const encryptPassword = (password: string) => {
-  const encryptedPassword = CryptoJS.AES.encrypt(password, secretKey).toString();
+  const encryptedPassword = CryptoJS.AES.encrypt(
+    password,
+    secretKey
+  ).toString();
   return encryptedPassword;
 };
 
 const decryptPassword = (encryptedPassword: string) => {
-  const decryptedPassword = CryptoJS.AES.decrypt(encryptedPassword, secretKey).toString(CryptoJS.enc.Utf8);
+  const decryptedPassword = CryptoJS.AES.decrypt(
+    encryptedPassword,
+    secretKey
+  ).toString(CryptoJS.enc.Utf8);
   return decryptedPassword;
 };
 
-export { encryptPassword, decryptPassword };
+export { decryptPassword, encryptPassword };
