@@ -30,7 +30,7 @@ async function handleUserRegister(req, res) {
       name: user.name,
     });
   } catch (err) {
-    console.log("User registration error:" , err);
+    console.log("User registration error:", err);
     return res.status(500).json({ message: "Server error occurred" });
   }
 }
@@ -43,50 +43,47 @@ async function handleUserLogin(req, res) {
       .status(400)
       .json({ message: "Please, provide Email and Password" });
   }
-   
-  try{
+
+  try {
     const user = await User.findOne({ email });
 
-  if (!user || !(await user.comparePassword(password))) {
-    return res.status(401).json({ message: "Invalid Username or Password" });
-  }
+    if (!user || !(await user.comparePassword(password))) {
+      return res.status(401).json({ message: "Invalid Username or Password" });
+    }
 
-  const token = generateToken(user);
-  // res.cookie("token", token);
-  return res.status(200).json({
-    message: "Logged in successfully",
-    token: token,
-    id: user._id,
-    email: user.email,
-    name: user.name,
-  });
-  }catch(err){
+    const token = generateToken(user);
+    // res.cookie("token", token);
+    return res.status(200).json({
+      message: "Logged in successfully",
+      token: token,
+      user: { id: user._id, email: user.email, name: user.name },
+    });
+  } catch (err) {
     console.log(err);
     return res.status(500).json({ message: "Server error occurred" });
   }
 }
 
 async function handleUserLogout(req, res) {
-  try{
+  try {
     res.clearCookie("token");
-  return res.status(200).json({ message: "Logged out successfully" });
-  }catch(err){
+    return res.status(200).json({ message: "Logged out successfully" });
+  } catch (err) {
     console.log(err);
     return res.status(500).json({ message: "Server error occurred" });
   }
 }
 
 async function handleUserProfile(req, res) {
-  try{
-      const user = await User.findById(req.user.id).select("name email");
-  return res
-    .status(200)
-    .json({ message: "Successfully fetched profile", user: user });
-  }catch(err){
+  try {
+    const user = await User.findById(req.user.id).select("name email");
+    return res
+      .status(200)
+      .json({ message: "Successfully fetched profile", user: user });
+  } catch (err) {
     console.log(err);
     return res.status(500).json({ message: "Server error occurred" });
   }
-
 }
 
 module.exports = {
