@@ -1,121 +1,72 @@
-import { API_URL, useAuth } from "@/app/context/AuthContext";
-import { decryptPassword, encryptPassword } from "@/utils/encryption";
-import axios from "axios";
-import { useEffect, useMemo, useState } from "react";
+import {
+  CreatePasswordDTO,
+  Password,
+  UpdatePasswordDTO,
+} from "@/types/password";
+import { passwordService } from "@/utils/passwordService";
+import { QueryClient, useMutation, useQuery } from "@tanstack/react-query";
 
-const usePasswords = () => {
-  const [passwords, setPasswords] = useState<Password[]>([]);
-  const [loading, setLoading] = useState<boolean>(false); // initialize to false for better control
-  const [error, setError] = useState<string | null>(null);
+// Define query client
+export const queryClient = new QueryClient();
 
-  const { authState } = useAuth();
+// Hook to fetch all passwords
+export const usePasswords = () => {
+  // Define query
+  const {
+    data: passwords = [],
+    error,
+    status,
+  } = useQuery<Password[]>({
+    queryKey: ["passwords"],
+    queryFn: passwordService.fetchPasswords,
+  });
 
-  // Memoize headers to prevent unnecessary re-renders
-  const headers = useMemo(() => {
-    return authState?.token ? { Authorization: `Bearer ${authState.token}` } : {};
-  }, [authState?.token]);
+  // Define mutation for adding password
+  const addPasswordMutation = useMutation<Password, Error, CreatePasswordDTO>({
+    mutationFn: passwordService.addPassword,
+    onSuccess: () => {
+      queryClient.invalidateQueries({
+        queryKey: ["passwords"],
+        refetchType: "all",
+      });
+    },
+  });
 
-  // Refactored fetchPasswords with better loading state management
-  const fetchPasswords = async () => {
-    setLoading(true);
-    setError(null);
-    
-    try {
-      const response = await axios.get(`${API_URL}/passwords`, { headers });
+  // Define mutation for deleting password
+  const deletePasswordMutation = useMutation<void, Error, string>({
+    mutationFn: passwordService.deletePassword,
+    onSuccess: () => {
+      queryClient.invalidateQueries({
+        queryKey: ["passwords"],
+        refetchType: "all",
+      });
+    },
+  });
 
-      const data = await response.data?.passwords || [];
-      const encryptedPasswords = data.map((password: Password) => ({
-        ...password,
-        password: decryptPassword(password.password),
-      }));
-
-      setPasswords(encryptedPasswords);
-    } catch (err) {
-      console.error("Error fetching passwords:", err);
-      setError("Failed to fetch passwords");
-    } finally {
-      setLoading(false);
+  // Define mutation for updating password
+  const updatePasswordMutation = useMutation<
+    Password,
+    Error,
+    {
+      id: string;
+      data: UpdatePasswordDTO;
     }
-  };
-
-  const addPassword = async (newPassword: Password) => {
-    setLoading(true);
-    setError(null);
-
-    try {
-      const result = await axios.post(
-        `${API_URL}/passwords`,
-        {
-          ...newPassword,
-          password: encryptPassword(newPassword.password),
-        },
-        {
-          headers,
-        }
-      );
-      fetchPasswords(); // refresh passwords list after addition
-
-      return result;
-    } catch (err) {
-      console.log("Failed to add password:", err);
-      setError("Failed to add password");
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  const updatePassword = async (id: string, updatedPassword: Password) => {
-    setLoading(true);
-    setError(null);
-    try {
-      const result = await axios.put(
-        `${API_URL}/passwords/${id}`,
-        {
-          ...updatedPassword,
-          password: encryptPassword(updatedPassword.password),
-        },
-        headers
-      );
-
-      fetchPasswords(); // refresh passwords list after update
-      return result
-    } catch (err) {
-      console.log("Failed to update password:", err);
-      setError("Failed to update password");
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  const deletePassword = async (id: string) => {
-    setLoading(true);
-    setError(null);
-
-    try {
-      await axios.delete(`${API_URL}/passwords/${id}`, { headers });
-      fetchPasswords(); // refresh passwords list after deletion
-    } catch (err) {
-      console.error("Failed to delete password:", err);
-      setError("Failed to delete password");
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  // Fetch passwords on component mount
-  useEffect(() => {
-    fetchPasswords();
-  }, []);
+  >({
+    mutationFn: ({ id, data }) => passwordService.updatePassword(id, data), // Function to update password
+    onSuccess: () => {
+      queryClient.invalidateQueries({
+        queryKey: ["passwords"],
+        refetchType: "all",
+      });
+    },
+  });
 
   return {
     passwords,
-    loading,
     error,
-    fetchPasswords,
-    addPassword,
-    updatePassword,
-    deletePassword,
+    status,
+    addPassword: addPasswordMutation.mutateAsync,
+    deletePassword: deletePasswordMutation.mutateAsync,
+    updatePassword: updatePasswordMutation.mutateAsync,
   };
 };
-
-export default usePasswords;
